maybe? fix tour crashes (ng)
i am not sure where it's crashing as i can't replicate it on debug but i just know it's here. at least i hope so... i was able to get through 3 rock power challenges like with this on wii so if this doesn't fix the crashes then idk lol we're fucked.