actually "fix"? tour crash (wii)
i have no clue why the fuck this is an issue or even what is going on i just know that r3 is somehow pointing to something very very early in memory. this fixes it but like what what the fuck i don't get it at all. anyways fuck the wii i love ppc <3